federator: implement dummy endpoints
